% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{AH_Tab12_1_1}
\alias{AH_Tab12_1_1}
\title{Annual maximum discharges of the Guadalupe River near Victoria, Texas, 1935-1978.}
\format{
A data frame with 44 rows and 1 variable:
\describe{
  \item{Q_cfs}{annual maximum discharges in cfs}
  ...
}
}
\source{
Chow, V. T., Maidment, D. R., & Mays, L. W. (1988). \emph{Applied Hydrology}.
   McGraw-Hill, New York, U. S.
}
\description{
A dataset containing annual maximum discharges (in cfs) of the Guadalupe River near Victoria, Texas,
   during 1935-1978, in cfs extracted from TABLE 12.1.1 of "Applied Hydrology" (Chow et al., 1987).
}
