% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{AH_Tab12_2_1}
\alias{AH_Tab12_2_1}
\title{Annual maximum 10-minute rainfall at Chicago, Illinois, 1913-1947.}
\format{
A data frame with 35 rows and 1 variable:
\describe{
  \item{PMax10min_in}{annual maximum 10-minute rainfall in inches}
  ...
}
}
\source{
Chow, V. T., Maidment, D. R., & Mays, L. W. (1988). Applied Hydrology.
   McGraw-Hill, New York, U. S.
}
\description{
A dataset containing annual maximum 10-minute rainfall (in inches) at Chicago, Illinois,
   during 1913-1947 extracted from TABLE 12.2.1 of "Applied Hydrology" (Chow et al., 1987).
}
