% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{B17C_Tab10_14}
\alias{B17C_Tab10_14}
\title{U.S. Geological Survey gage 07099500 (and others) Arkansas River annual peak-flow record during 1864-1976}
\format{
A data frame with 85 rows and 1 variable:
\describe{
  \item{Q_peak_cfs}{peak flows in cfs}
}
}
\source{
England, J. F., Jr., Cohn, T. A., Faber, B. A., Stedinger, J. R., Thomas, W. O., Jr.,
   Veilleux, A. G., Kiang, J. E., & Mason, R. R., Jr. (2019).
   Guidelines for determining flood flow frequency - Bulletin 17C. U.S. Geological Survey.
}
\description{
A dataset containing the U.S. Geological Survey gage 07099500 (and others) Arkansas River annual peak-flow
   record consisting of 85 peaks from 1864 to 1976 extracted from Table 10.14 in "Guidelines for determining
   flood flow frequency - Bulletin 17C" (England et al., 2019)
}
\details{
This table contains the water year of the annual peak and the corresponding annual peak in
   cubic feet per second (ft3/s).
}
