% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{B17C_Tab10_2}
\alias{B17C_Tab10_2}
\title{U.S. Geological Survey gage 01134500 Moose River at Victory, Vermont annual peak-flow record during 1947-2014}
\format{
A data frame with 68 rows and 1 variable:
\describe{
  \item{Q_peak_cfs}{peak flows in cfs}
}
}
\source{
England, J. F., Jr., Cohn, T. A., Faber, B. A., Stedinger, J. R., Thomas, W. O., Jr.,
   Veilleux, A. G., Kiang, J. E., & Mason, R. R., Jr. (2019).
   Guidelines for determining flood flow frequency - Bulletin 17C. U.S. Geological Survey.
}
\description{
A dataset containing the U.S. Geological Survey gage 01134500 Moose River at Victory, Vermont annual
   peak-flow record consisting of 68 peaks from 1947 to 2014 extracted from Table 10.2 in "Guidelines for
   determining flood flow frequency - Bulletin 17C" (England et al., 2019)
}
\details{
This table contains the date of the annual peak recorded at the gage, the water year of the annual peak,
   and the corresponding annual peak in cubic feet per second (ft3/s).
}
