% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{B17C_Tab10_6}
\alias{B17C_Tab10_6}
\title{U.S. Geological Survey gage 11274500 Orestimba Creek near Newman, California annual peak-flow record
   during 1932-2013}
\format{
A data frame with 82 rows and 1 variable:
\describe{
  \item{Q_peak_cfs}{peak flows in cfs}
}
}
\source{
England, J. F., Jr., Cohn, T. A., Faber, B. A., Stedinger, J. R., Thomas, W. O., Jr.,
   Veilleux, A. G., Kiang, J. E., & Mason, R. R., Jr. (2019).
   Guidelines for determining flood flow frequency - Bulletin 17C. U.S. Geological Survey.
}
\description{
A dataset containing the U.S. Geological Survey gage 11274500 Orestimba Creek near Newman, California
   annual peak-flow record consisting of 82 peaks from 1932 to 2013 extracted from Table 10.6 in "Guidelines for
   determining flood flow frequency - Bulletin 17C" (England et al., 2019)
}
\details{
This table contains the date of the annual peak recorded at the gage, the water year of the annual peak,
   and the corresponding annual peak in cubic feet per second (ft3/s).
}
