% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{B17C_Tab8_1}
\alias{B17C_Tab8_1}
\title{Observed annual peak data for the Etowah River and Suwanee Creek from 1985-2004}
\format{
A data frame with 20 rows and 2 variables:
\describe{
  \item{Q_peak_Etowa_River_cfs}{name of the crop}
  \item{Q_peak_Suwanee_Creek_cfs}{the crop coefficient in the growth initial stage}
}
}
\source{
England, J. F., Jr., Cohn, T. A., Faber, B. A., Stedinger, J. R., Thomas, W. O., Jr.,
   Veilleux, A. G., Kiang, J. E., & Mason, R. R., Jr. (2019).
   Guidelines for determining flood flow frequency - Bulletin 17C. U.S. Geological Survey.
}
\description{
A dataset containing the summary of concurrent observed annual peak data for the Etowah River and
   Suwanee Creek from 1985-2004 extracted from Table 8.1 in "Guidelines for determining flood flow
   frequency - Bulletin 17C" (England et al., 2019)
}
