% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{B17C_Tab8_2}
\alias{B17C_Tab8_2}
\title{MOVE extended record for 13 years (1972-1984) for Suwanee Creek at Suwanee, Georgia}
\format{
A data frame with 13 rows and 1 variable:
\describe{
  \item{Q_peak_cfs}{peak flows in cfs}
}
}
\source{
England, J. F., Jr., Cohn, T. A., Faber, B. A., Stedinger, J. R., Thomas, W. O., Jr.,
   Veilleux, A. G., Kiang, J. E., & Mason, R. R., Jr. (2019).
   Guidelines for determining flood flow frequency - Bulletin 17C. U.S. Geological Survey.
}
\description{
A dataset containing the MOVE extended record for 13 years (1972-1984) for Suwanee Creek at Suwanee,
   Georgia (station 02334885) extracted from Table 8.2 in "Guidelines for determining flood flow
   frequency - Bulletin 17C" (England et al., 2019)
}
