% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{B17C_Tab8_3}
\alias{B17C_Tab8_3}
\title{Flood records for 93 years (1892-1984) for the Etowah River at Canton, Georgia}
\format{
A data frame with 93 rows and 1 variable:
\describe{
  \item{Q_peak_cfs}{peak flows in cfs}
}
}
\source{
England, J. F., Jr., Cohn, T. A., Faber, B. A., Stedinger, J. R., Thomas, W. O., Jr.,
   Veilleux, A. G., Kiang, J. E., & Mason, R. R., Jr. (2019).
   Guidelines for determining flood flow frequency - Bulletin 17C. U.S. Geological Survey.
}
\description{
A dataset containing the flood records for 93 years (1892-1984) for the Etowah River at Canton,
   Georgia (station 02335000) extracted from Table 8.3 in "Guidelines for determining flood flow
   frequency - Bulletin 17C" (England et al., 2019)
}
