% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FloodFreqPlot.R
\docType{package}
\name{FloodFreqPlot-package}
\alias{FloodFreqPlot-package}
\title{A package for flood frequency analysis by a graphical method}
\description{
The flood data are plotted on an appropriate probability paper that linearizes the cumulative distribution function.
   Then the plotted flood data are fitted with a straight line for interpolation and extrapolation purposes.
}
