% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{Harricana}
\alias{Harricana}
\title{Maximum annual peak discharge values, obseved at Harricana River at Amos}
\format{
A data frame with 72 rows and 1 variable:
\describe{
  \item{Observation}{peak flows}
}
}
\source{
Bobee, B. & Ashkar, F. (1991). \emph{The Gamma Family and Derived Distributions Applied in Hydrology}.
   Water Resources Publications.
}
\description{
A dataset containing the Maximum annual peak discharge values in cubic meter per second (cms), observed at
   Harricana River at Amos (Quebec, Canada) as displayed by the program HFA extracted from Table 1.2 in
   "The Gamma Family and Derived Distributions Applied in Hydrology" (Bobee and Ashkar, 1991).
}
